+++ folio-spring-i18n/src/test/java/org/folio/spring/i18n/model/TranslationFileLanguageExtractionTest.java
@@ -9,12 +9,14 @@ import static org.junit.jupiter.params.provider.Arguments.arguments;
 
 import java.util.Arrays;
 import java.util.List;
+import org.folio.spring.testing.type.UnitTest;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.core.io.FileSystemResource;
 
+@UnitTest
 class TranslationFileLanguageExtractionTest {
 
   static List<Arguments> fullExtractionCases() {